update
데이터 통합, 예약 수정, ResrvationDTO 구현

+++ KNUCinema/src/main/java/com/example/KNUCinema/ReservationDTO.java
@@ -8,7 +8,9 @@ import lombok.*;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ReservationDTO {
-    private int id;
+    private long id;
     private CinemaDTO cinema;
     private int UserId;
 }
+
+//TODO : id는 자동으로 set금지 하는 어노테이션
